Move ChannelConnectivityObserver to its own file
