+++ Sources/SwiftGRPC/Core/Channel.swift
@@ -15,7 +15,6 @@
  */
 #if SWIFT_PACKAGE
 import CgRPC
-import Dispatch
 #endif
 import Foundation
 
@@ -126,81 +125,5 @@ public class Channel {
   /// - Parameter callback: block executed every time a new connectivity state is detected
   public func addConnectivityObserver(callback: @escaping (ConnectivityState) -> Void) {
     connectivityObserver.addConnectivityObserver(callback: callback)
-  }
-}
-
-private extension Channel {
-  final class ConnectivityObserver {
-    private let completionQueue: CompletionQueue
-    private let underlyingChannel: UnsafeMutableRawPointer
-    private let underlyingCompletionQueue: UnsafeMutableRawPointer
-    private var callbacks = [(ConnectivityState) -> Void]()
-    private var hasBeenShutdown = false
-    private let stateMutex = Mutex()
-
-    init(underlyingChannel: UnsafeMutableRawPointer) {
-      self.underlyingChannel = underlyingChannel
-      self.underlyingCompletionQueue = cgrpc_completion_queue_create_for_next()
-      self.completionQueue = CompletionQueue(underlyingCompletionQueue: self.underlyingCompletionQueue,
-                                             name: "Connectivity State")
-      self.run()
-    }
-
-    deinit {
-      shutdown()
-    }
-
-    private func run() {
-      let spinloopThreadQueue = DispatchQueue(label: "SwiftGRPC.ConnectivityObserver.run.spinloopThread")
-      var lastState = ConnectivityState(cgrpc_channel_check_connectivity_state(self.underlyingChannel, 0))
-      spinloopThreadQueue.async {
-        while true  {
-          guard (self.stateMutex.synchronize { !self.hasBeenShutdown }) else {
-            return
-          }
-
-          guard let underlyingState = lastState.underlyingState else { return }
-
-          let deadline: TimeInterval = 0.2
-          cgrpc_channel_watch_connectivity_state(self.underlyingChannel, self.underlyingCompletionQueue,
-                                                 underlyingState, deadline, nil)
-
-          let event = self.completionQueue.wait(timeout: deadline)
-          guard (self.stateMutex.synchronize{ !self.hasBeenShutdown }) else {
-            return
-          }
-
-          switch event.type {
-          case .complete:
-            let newState = ConnectivityState(cgrpc_channel_check_connectivity_state(self.underlyingChannel, 0))
-            guard newState != lastState else { continue }
-
-            lastState = newState
-            self.stateMutex.synchronize {
-              self.callbacks.forEach { callback in callback(newState) }
-            }
-
-          case .queueShutdown:
-            return
-
-          default:
-            continue
-          }
-        }
-      }
-    }
-
-    func addConnectivityObserver(callback: @escaping (ConnectivityState) -> Void) {
-      self.stateMutex.synchronize {
-        self.callbacks.append(callback)
-      }
-    }
-
-    func shutdown() {
-      stateMutex.synchronize {
-        hasBeenShutdown = true
-      }
-      completionQueue.shutdown()
-    }
   }
 }
